# introduce `unsafe_proto`