@kzc 2017-02-16
I'm not in favor of this optimization because it generates a temporary object in each case that may produce slower code:
```js
        Array.prototype.splice.apply(a, [1, 2, b, c]);
        Object.prototype.hasOwnProperty.call(d, "foo");
        String.prototype.indexOf.call(e, "bar");
```
--->
```js
        [].splice.apply(a, [1, 2, b, c]);
        ({}).hasOwnProperty.call(d, "foo");
        "".indexOf.call(e, "bar");
```
Maybe if it were under a different option other than `unsafe`.

If you can demonstrate in micro-benchmarks on Chrome, FireFox, IE, Edge, Webkit/Safari that it is of equivalent speed to the original, I could be persuaded to keep it under `unsafe`.