@kzc 2017-02-16
`unsafe_prototype`?