@alexlamsl 2017-02-16
[Adjustment of whitespaces](https://github.com/mishoo/UglifyJS2/blob/0610c020b1544820be9898a285ab6c9066490552/lib/compress.js#L51-L81) would be dramatic 😅