@alexlamsl 2017-02-16
@kzc as for https://github.com/mishoo/UglifyJS2/pull/1491#issuecomment-280339543, I've found the following _jsperf_ links that might be relevant:

https://jsperf.com/array-prototype-slice-call-vs-slice-call/17
https://jsperf.com/foreach-vs-array-prototype-foreach