@kzc 2017-02-16
https://jsperf.com/prototype-call-vs-literal-call
```
Chrome 56:   5% slower with unsafe_proto
FireFox 51:  no difference
Safari 9:    10% slower with unsafe_proto
```