@alexlamsl 2017-02-16
`IE 11:   2% slower with unsafe_proto`