@kzc 2017-02-16
The `unsafe_proto` optimization is not bad on modern JS engines, but by default uglify shouldn't produce code that's slower than unminified code.

Would be curious if you could run it on older browsers - IE8/9/10. I don't have any handy.