@alexlamsl 2017-02-16
```
Opera 12: 15% slower with unsafe_proto
IE 8:     typeof Date.now === "undefined"
Safari 5: 10% slower with unsafe_proto
FF 43:    no difference
```