@kzc 2017-02-16
IE8 is useless. Otherwise much like the timings for modern engines.

Are you still okay with using the flag `unsafe_proto`?