@alexlamsl 2017-02-16
I'm happy with the extra flag either way - extra granularity is never an issue for me.